requirements.txt
torch
diffusers
transformers
streamlit == 1.40.0
streamlit-drawable-canvas
numpy
pillow